// src/interface/weeklyreport/type.ts
export interface IGetAttendanceDetails {
  params: IGetAttendanceDetailsParams;
}

export interface IGetAttendanceDetailsParams {
  args: number[];
}
// ...
export interface InvalidTimesheet {
  last_attendance_attempt: string;
  first_attendance_attempt: string;
  attendance_missing_from: string;
  attendance_missing_to: string;
  id: number;
  reason: boolean;
  remarks: string;
  validated: string;
  employee_id: number;
  employee_name: string;
  employee_code: string;
  department: string;
  position: boolean;
  invalid_date: string;
  invalid_type: string;
  shift_from: string;
  shift_to: string;
  shift_break: boolean;
  real_time_attendance_data: string;
  validation_data: string;
  content: boolean;
  create_date: string;
  keep_in_time: boolean;
}

export interface IListLeavereq {
  id: number;
  created_date: string;
  from_date: string;
  holiday_status_id: HolidayStatusId;
  state: string;
  employee_id: EmployeeId;
  date_to: string;
  config_id: ConfigId;
  for_reasons: string;
  reasons: string;
  time: number;
  minutes: number;
  overtime_from: any;
  overtime_to: any;
  attendance_missing_from: any;
  attendance_missing_to: any;
  convert_overtime: boolean;
  multiplier_wage: number | null;
  multiplier_work_time: number | null;
  employee_parent_id: number | null;
  hr_approval_id: number | null;
  work_trip_location: number | null;
  holiday_request_url_ids?: Array<any>;
  locked:any
}

export interface ConfigId {
  id: number | boolean;
  name: string | boolean;
}

export interface HolidayStatusId {
  id: number;
  name: string;
  type: string;
}

export interface EmployeeId {
  id: number;
  name: string;
}

export interface historyShiftEditBody {
  params: historyShiftEditParams;
}
export interface historyShiftEditParams {
  args: string[];
}